It's possible to authenticate through the Admin API.
Note: Had to add more dependencies because Keystone returns extant
tokens, so we have to be very careful about ordering of token
revocations.  This shouldn't affect anything that extends
base.KeystoneTest--all that can still run in parallel--but the
test_adminauth() has to be a dependency of
base.KeystoneTest.setUpClass() and has to be dependent on
test_authenticate() so we serialize things properly.