Cleaned up come issues with python2.6.
While working on packaging, we've discovered that the unittests do not work
on python2.6/Ubuntu Maverick. The first change to
keystone/common/template.py fixes an exception misuse. The rest are all
being explicit about using unittest2, because on 2.6 in some cases if you
just import unittest, you get left with an old API.

Change-Id: I783ef1cc14860578f8f61ae23e1afe8ac8f293a7