Bug #854104   - Changes to allow admin url to be shown only for admin users.   - Additional test asserts to verify.
This was implemented by adding a parameter to the AuthData __init__
(the class that renders auth call responses). The url_types parameter
filters which types of URLs to return; admin, intenrnal, public.
And in logic/service.py, there is authorization logic which checks
the caller's rights and based on that determines if they can see
the 'admin' URL. If not, it takes it out of the array passed to
the url_types paramater passed to the AuthData initializer.

Added method documentation and renamed a couple in service.py in
the process to improve code understandability. More needs to be
done, but moving the ball forward.

Change-Id: I3fb90c31617da7647a211901d647cfb53f182be2