Fixes a number of configuration/startup bugs
LP Bug#897376 - log_file in configuration file not respected

Adds a check to ensure that the log_file in the configuration
file, if set, is not overridden by the default options['log_file']
value set in keystone.common.configu.load_paste_app()

LP Bug#897382 - service_ssl not safely checked for existence

Added safe checks for a number of SSL-related variables in
bin/keystone to help support easy migration for existing
configuration files that did not have the SSL options in them.

LP Bug#897397 - bin/keystone-auth does not respect service_port
                or service_host

Replicates the same behaviour that is in keystone-admin into
keystone-auth so that it recognizes service_host and service_port
instead of only bind_host and bind_port. This enables you to
pass the main keystone.conf file to keystone-auth and it will
not bomb.

Note: keystone.common.config should go away once openstack.common.config
is a reality...

Change-Id: If2dfa57ba00758144219f8c1d42c05e56ed44ca2