Bug#899116: use correct module when building docs
Currently the wrong directory is searched for the
keystone module, requiring keystone to be installed
in the default system path, for docs to be buildable.
Also the path was appended, so the default system version
which may be completely different is referenced.
This ensures that only local keystone files are referenced.

Change-Id: I187a5153d9c8be076e8aff6bb0ae63a7c739e3e8