Fixed version response (bug 891555 and bug 843052)
891555: Version response was supposed to be wrapped in a
'versions' element, instead there was only one version
being returned at the root.
Fixed and added tests to validate that.

843052: requests for version ATOM were ignoring the
atom+xml as a valid content-type. Fix required
adding a new atom template file and adding a new
mapping for the content-type/extension in the
URL normalizer.
Fixed and added tests to validate that.

Change-Id: I259c060e592e11cfc06f59ac5a540178c7f2def2