Fixed error with database initialization
New versioning code was not creating tables in
database when database was not an in-memory database.
This broke existing behaviour, scripts (like
devstack).
Identified problem in backend tests where backends
were not being reloaded between tests.
This highlighted an issue where LDAP was not passing
tests and that was fixed.

Change-Id: Ib4eab7fe696bdc48c363032f078e5c535abc0849