Move URL Normalizer to Frontends
keystone.middleware.url is a front-end filter for
keystone. It is not a service middleware. It should
therefore not be in the middleware directory.

This patch moves it and updates associated documentation
and tests. It also maintains a reference file and code
to allow existing configurations to continue to work.
They will receive a warning message in the log to
update their configuration.

Change-Id: Ib946bd3883df1cf5855b207771669f91d0dad1da