Remove install_requires processing.
This is based on a similar change to nova:
https://review.openstack.org/#change,2800

Loading install_requires with the contents of pip-requires
isn't getting us any real beneift and is causing issues.

a) It can conflict with installing keystone into an environment
where deps have been installed from packages (devstack)
b) It breaks the ability to use -e git urls in pip-requires
which we want to start using for python-novaclient and
python-keystoneclient
c) It causes spurious network traffic when we're trying to
test things.

At the same time, since we are not expecting anyone to install
keystone from setup.py for production, the normal benefit of the
feature is not needed.

Change-Id: I7235e1bff75a9fd1455daa684300a83c4295bd9c