Use gerrit instead of github
When we run gating tests in jenkins, any access of network resources is a
potential source of false-negative on the test due to intermittent service
failures on systems that are out of our control. We observe that this is
actually quite frequent when things want to access PyPI or github. With
pypi, we pre-create virtualenvs and cache the eggs so that an individual
test run doesn't fail due to pypi not responding. For repos, if at all
possible, we direct them all at the gerrit instance, because since gerrit is
driving the test run in the first place, it's indicative of a much larger
problem if jenkins can't talk to it - and it's one that we can fix if it
does come up.

Change-Id: I9f54133f7f2025d15a9d0b270d2466438cbc6dd5