Improve auth_str_equal().
This patch is to improve auth_str_equal() a bit.  The whole point of
this function is to do a string comparison in constant time to help
protect against timing attacks.  The original implementation had a bit
of a silly property in that it would exit early if the strings were not
of the same length.  This would theoretically still allow someone to
discover the proper length of a password.

This patch moves the length verification to the end.  It also makes it
so the main loop time to run is a function of the provided password
length instead of the length of the shorter of the two strings.

Change-Id: I6dbe076818b7e3e8a313544ebd5c5734b5a025e5