Support PyPAM in pam backend, update to latest API
Fix bug 938801.

This bug pointed out that some distros don't have the same pam Python
module packaged that this backend was expecting.  In my case, on Fedora,
it's PAM and the API is not compatible with the pam module that was
used.  This patch makes the backend support PyPAM, as well as the
original pam module that was used.

In order to test this, I updated the pam backend to the latest backend
API.  Even though the base class will raise NotImplementedError, I
included all functions here to make it more clear all of the things
this backend does not do.

Change-Id: I74144f4e63b6830c8224bc87e1662eb5df8728a0