notify calling process we are ready to serve
Fixes bug 980037

Service managers starting keystone-all have no way of being notified
when the service is ready to accept connections. This commit allows
a configurable command to be called when we are ready e.g.
for systemd setting the statup type of a service unit to "notify" and setting
onready    = systemd-notify --ready
in keystone.conf, would notify a waiting systemd that we are ready to
serve

In an automated envirnment (e.g. puppet) this will allow the startup of
the keystone-all service (with systemctl for example) directly followed
by usage of the keystone client without the need for a sleep (or retry)
while we are waiting for the keystone service to be ready.

Change-Id: I3f7aafe9837be60a0f35cae1a7db892f6851cc47