Fix test env for the stable branch
Need both changes in one commit to pass the gate!

* Nail pep8 dependencies to 1.0.1.

Nails the pep8 deps for tox and test-requires to 1.0.1.
Fixes an issues causing pep8 failures due to a new pep8 release.

(cherry picked from Nova stable)

* Switch to 1000 rounds during unit tests

Fixes bug 992918

passlib 1.6 introduced a minimum number of rounds for sha512_crypt. As
a result, increase the rounds used during testing to the minimum

Change-Id: Ic0c635e92b4f13180a047904a6efa490ab599012