Simplify the sql backend deletion of users and tenants.
There is a remaining problem in that the table definition permits dangling
membership in tenants, and vice verca, but this change will make the fix for
that easier to review, and make the code simpler and faster at the sametime.

See bug 1000609 for the bug report that lead to examining this.

Change-Id: Id7cd5fad7032779d352a7c577c8d10558091d767