Fixes Bug 1063852
Add in the issue time, to prevent a race condition where a token is issued and
revoked, and then a request for an additional token is processed identical
to the first.  Each token now contains the issue time to make it unique.

(moving changing default to PKI to separate review)

Change-Id: I26ed5b3bb31840f5baaf64dbcbeac477e4d71afd