bug 1068674
Redo part of bp/sql-identiy-pam that was accidently undone by bug 968519.

We encapsulated the call to utils.check_password with a local method,
_check_password, to make it easier to subclass Identity. This allows us
to use a different password checker without having to replace the entire
authenticate method in our code.  The fix for 968519 accidently removed
the call to the local method. *This* fix puts that call back in.

Updating comment because Jenkins failed due to build timeout in
unrelated test.

Change-Id: I69a3ba2d5a62e4c600edab7ef2cc07413c7360cc