Raise exception if openssl stderr indicates one.
The cms module calls openssl and raises a CalledProcessError if the
return code is non zero. However, I've observed issues where the return
code is zero but the std err shows the command failed. This commit
changes the test to also look for the text "Error" in openssl's stderr
and raise CalledProcessError in that event as well.

It also removes the keyword arg "output" to CalledProcessError as it
wasn't being found.

Change-Id: I2b2a9c026557632d0c71dd6f987eaa263f387242