Make tox.ini run pep8 checks on bin.
This updates the pep8 checks in our tox.ini file so that
we are also scanning the bin directory.

Additionally, it updates the main pep8 check so that it scans
keystone properly as well. Previously there were a bunch of files
getting skipped due to some of the pep8 pattern matching.

Change-Id: I13827f1c1e4155aa6979c10e981da5422391bf2d