Reduce total number of fixtures
Fixtures are created before every test, so each fixture adds a
considerable amount of overhead to the overall test suite.

This patch attempts to eliminate fixtures utilized by only a few tests
in favor of re-cycling as many fixtures as possible. As a result, a few
tests are refactored to depend on different fixtures.

Change-Id: Idd4dcef5e38e304d19110c61886887fb64b4d658