Adding a means to connect back to a pydevd debugger.
That patch allows a developer to remotely run a pydev debugger and have the
keystone-all process connect back to it.  Two command line options are
introduced:
    --pydev-debug-host <host>
    --pydev-debug-port <port>
both of the above options are required to enable this behavior.

This patch only enables this behavior when the service is started with
keystone-all.  In the future parts of this patch can be used to enable
this behavior when running in Apache.

Change-Id: I92f99fa34112336a96e42e8261b7313f23ee994e