shorten pep8 output
This removes the PEP8 output that describes in detail how to fix each
problem. It makes pep8 review so long that the earliest lines scroll
out of the scrollback buffer before they can even be read.

With this change, each violation is on line long, which is
much more readable.

Change-Id: I0d8cc64fd6027419754732e314c047b3775a121c