Test 0.2.0 keystoneclient to avoid new deps
As described here:

  http://lists.openstack.org/pipermail/openstack-dev/2013-January/004240.html

keystone tests check out various different versions of keystoneclient
and uses them to run a set of tests against the keystone API.

A recent change to keystoneclient in version 0.2.1 introduced a new
dependency on the requests module which is not currently listed as
a dependency for keystone's tests.

Rather than chasing latest keystoneclient on stable/folsom, let's
just test the 0.2.0 version.

Change-Id: I730f0d7e227c76ae77fdb012feec9ce41ab8d8d2