rename tenant to project in sql
Change is motivated by the need to do an incremental conversion from using
tenant to using project as the name.  The database is isolated from the API
and can be modified without breaking integrations.  ADditional work wil be required
to get the API to meet the V3 spec without breaking the V2 spec

Change-Id: I7cf7695354071f0ea6252be4730ceec3af0a2f35