Ensure keystone unittests do not leave CONF.policyfile in bad state
A few of the tests overwrite the name of the policy file (so they can
use a temp file instead).  However, if it is left that way, subsequent
tests that rely on it may fail.

A separate patch will look at doing a more comprehensive reset in the
setup of test_v3 - ensuring we always start from a completely clean
slate for all confirguration parameters.

Fixes Bug #1131819

Change-Id: Ibe5ee12f44310de00b12ddd405c83f59b2d840b7