Unpin pam dependency version
Fixes bug #1133240

keystone has always pinned the pam dependency to 0.1.4

  pam>=0.1.4

this version was released in November 2009 and there hasn't been a
release since. I can't even find a version control repo for the
project.

There's no particular reason to expect a future version would break us
and, indeed, we want to find out about it if it happens.

Change-Id: I2becf47c9d2566e2c3496ae15ca886b737006a92