Migrate roles from metadata to user_project_metadata
migration 017 did not migrate existing roles from the metadata table
that was created in 001. Adding a migration (20) that compares the roles
in the metadata table (if any) and joins them to the new
user_project_metadata role that matches the user and tenant from the old
table. Also adding subsequent tests to check both of the issues above.
bug 1131087

Change-Id: I00ea6043d949c9c358827e25f05c63515fe5dea8