Discard null endpoints (bug 1152632)
If a v2 client passed {..., "adminurl": null, ...} in a endpoint-create
request, then the null value was being persisted on an endpoint with a
different interface value (i.e. a publicly facing endpoint would have an
"adminurl": null value inexplicably attached to it.)

This change simply pops null urls from the endpoint and discards them.

Change-Id: Idd0964b6ec34fbc8b979253d32f655ea9797f259