Skip IPv6 tests for eventlet dns
Eventlet address resolving doesn't support IPv6. This is causing the
gate to fail.

The failure happens when the dnspython package is installed in the
environment. When dnspython is there then eventlet takes over
getaddrinfo, and eventlet's getaddrinfo doesn't support IPv6. If
dnspython isn't there, then the eventlet patching doesn't happen
and IPv6 address resolution works.

It appears that the pip dependencies have recently been updated
so that dnspython gets installed now.

The fix is to skip the 3 tests that force the server to listen
on "::1".

Fixes bug 1176204

Once eventlet is updated, we can add the tests back in. Here's the
eventlet code in question:

https://bitbucket.org/eventlet/eventlet/src/e0f578180d7d82d2ed3d8a96d520103503c524ec/eventlet/support/greendns.py?at=0.12#cl-166

Change-Id: Ib7a5a5f0dcac82229b7b7fb3df4b1ef141586c5f