Disable eventlet monkey-patching of DNS
This change avoids eventlet's monkey-patching of DNS resolution.
eventlet's doesn't support IPv6, for example.

The way to avoid eventlet's DNS is to set an environment
variable. The trick is the environment variable needs to be set
before eventlet is imported.

A similar change was made in nova, so this is just copying that
code and technique to Keystone.

This allows re-enabling the IPv6 tests, too.

Change-Id: I1e4d4bbfea374d386796d1505a602fd187e75c67