Ec2 credentials table not created during testing
This change makes it so that the SQL Ec2Credential unit tests
can be run individually. It does this by making sure the
credential table model is available before the tables are
created.

Before this change, you couldn't run these tests individually
and would get an error that the ec2_credential table doesn't
exist.

Part of fix for bug 1179259

Change-Id: I68c8b91b18ac0065628c979e5bcc56152ae14916